remove reloc.h header.

+++ host/linux/windows.h
@@ -20,7 +20,7 @@ typedef bool BOOL;
 typedef uint8_t UINT8;
 typedef uint16_t UINT16;
 typedef uint32_t UINT32;
-typedef uint32_t UINT64;
+typedef uint64_t UINT64;
 typedef uint64_t UINTN;
 typedef void* LPVOID;
 typedef void* HMODULE;
